pyproject.toml
[project]
name = "check"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ddgs>=9.2.3",
    "openai-agents>=0.2.2",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.0",
    "twilio>=9.6.5",
]
